[Bronze V] Title: 이상한 기호, Time: 128 ms, Memory: 14288 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -12,8 +12,8 @@ public class Main {
 		
 		long result = 0l;
 		
-		int a = Integer.parseInt(st.nextToken());
-		int b = Integer.parseInt(st.nextToken());
+		long a = Long.parseLong(st.nextToken());
+		long b = Long.parseLong(st.nextToken());
 		
 		result = (a+b) * (a-b);
